ntrip_daemon fixes for unreliable connections (#343)
* ntrip_daemon: robustness improvements

+ Add more logging to ntrip_daemon / libnetwork, fix the logging setup
so that it follows the conventions of other daemons.

+ Fix a bug in the stall detection, incrementing the stall count needed
to happen separately from reseting the stall count (previously, if we
detected a stall, the count would be incremented, but then immediatedly
reset to zero since it wasn't greater than 30).

+ Make ntrip_daemon able to detect some existential threats... like the
pipe filling up and configure the receive buffer to be smaller so that
the receive window that's negotiated with the server is relatively
small.  NTrip sends about 1k of data per 1s pulse, so we shouldn't need
the receive buffer to be that large.

+ Sending SIGUSR1 to ntrip_daemon will cause it to reconnect to the
ntrip server: this is a last ditch effort to try to fix the connection
and get good base observation data again.  The sbp_rtcm3_bridge daemon
invokes this when the gnss-convertors lib notices that observation data
has become excessively old.