Sample app updates
+ Update post_image.sh to log the location of the build images, also
update the script so that it's less noisy, exporting DEBUG to the
environment will restore the prior logging behavior.

+ Move the S83sample_daemon start script into the package directory so
it can be conditionally installed if the sample daemon is enabled.

+ Change monit to include all .monitrc files in the /etc/monitrc.d
directory.

+ Add awscli to the make docker image so that `make firmware` can be run
inside the docker image (this will lessen the number of things that need
to be installed in VirtualBox for SDK users).

+ Attempt to fix the bug where the monit control port would not start
during boot if the eth0 interface was not up yet... this required
bumping to monit version 5.21.0 (which required forking buildroot)--
there may be a simpler workaround for this by using a unix socket.

+ Fix bug in docker base image build script

+ Fix bug in docker-wipe target

+ Remove sbp_router.yml changes, instead we'll rebroadcast POS and
HEARTBEAT messages over UDP.