Move the 'fetching firmware...' section to de-emphasize
The requirements of the 'fetching firmware' section is covered in the
building section (running the `firmware` or `docker-make-firmware`
targets).

Move the section down to de-emphasize the content, most consuming of the
SDK should not have to worry about this, and if it's an issue for
developers, we should look at the build environement to detect incorrect
configurations automatically (for example, if it's a non-release branch
then fail to start if there are no AWS credentials configured).