Add Warning when NTRIP Connection Ends with 200
If an ntrip connection finishes with response code
and no data received we assume that the endpoint
was expecting a GGA message in order to begin the
transmission of VRS corrections. This does not
cover the case where an RTCM message indicating
such a requirement is sent.

Summary of features added:

 * A response code data member and check function

 * The check function is executed on completion
   of a network request that did not have a curl
   error and had a non-zero response code

 * An implementation of this feature to warn when
   the ntrip download ends with 200 and no data

 * The warning will only be issued once as it is
   a large message and the network request may
   repeat many times with the same result

Addresses swift-nav/piksi_v3_bug_tracking#1012