ntpd: Use ZDA for time, and trust GPS date.
Update the mode for ntpd's NMEA refclock.
Bit 3 indicates that the ZDA message should be used for the time,
since this has complete date information.
Bit 25 is undocumented, but configured ntpd to trust the date
supplied by the receiver.  Without this ntpd assumes the
receiver doesn't handle week rollover correctly, and will select
the GPS epoch based on the system time (which is initially 1970).