Enable zync_remoteproc module, make remoteproc able to find fw
+ Remote proc apparently change how fw is loaded, you can echo a
firmware name into a sysfs entry or it has a default name that it uses
when looking in /lib/firmware.

This still doesn't allow the remoteproc to start, but it at least lets
it load fw.

+ Attempt to log a little more information from upgrade_tool-- which
fails around 88% when attempting to upgrade after flashing this image.
Not yet sure if this is a board problem or if there's an issue with
Linux.