Refactor of libpiksi settings to allow for watchers
The bulk of this refactor is to break the previously
large routine `setting_register` into multiple
setps that allow one to create a setting independent
of registering it with the settings daemon. This
allows for the reuse of the settings context, with
some modifications to internal data, to store
entries that only look for read responses as a
form of settings watch. The internal API
`settings_add_setting` serves the previous API
for registering settings as well as a new API
`settings_add_watch` that configures the new
setting entry to have watch only properties.

Notes:
 * Formatting and parsing have be factored into
   independent functions for reuse
 * Read response and write response callbacks
   are now added to zloop dynamically based
   on usage to reduce unnecessary message
   processing
 * Setting add routine prevents duplicates