Dump serial output buffer when it's full
At 57600 baud with a 20Hz solution frequency, Piksi produces a stream of
data that's about 5-7 KiB/s -- near the limit that the output queue can be
emptied at 57600 baud.

This causing settings reads (which are lots of small packets) to be lost
or arrive in between lots of solution messages.  The best thing we can
do here (short of implementing priority queuing and smart rate limiting)
is to issue a warning that we're getting behind.