Add user account for ntrip_daemon
Move ntrip settings out of piksi_system_daemon and into the
ntrip_daemon.

Add a unix domain socket that can be used to control the ntrip_daemon,
and require that it refresh it's connection (TODO: this is stubbed out
in piksi_system_daemon need to fill this in for use by sbp_rtcm3_bridge
and the cell modem wizard).

Setup permissions such that other processes can write the control socket
but not delete it.

Fixed a bug in sbp_zmq_pubsub wrt tearing down.