Add an explicit sync call to copy_sys_logs
This ensures that data actually makes it to the disk in the event of a
power failure.  For DEVC-585 we were seeing that a new directory was not
getting created in the event of a power cycle.  A new directory was
getting created, it was just the same as the one from the previous boot
because the data was never sync'd to the device.