Fix crash when dumping pppd output
Remove hack that tried to fix-up strings that pppd/chat would write one
character at a time... just in case this was contributing to the crash.

Also, guard against a null buffer pointer and make sure that we call
strlen on something that's guaranteed to be null temrinated.

The current working theory is that when the pppd daemon would
occasionally start outputing "garbage" from the ppp stream, this would
overflow the SBP max payload and strlen would not terminate the
buffer... thus the call to `strlen(storage)` would crash.