Do work to make sure pppd and chat are fully terminated
Add a wrapper for the chatscript that records the process id, so we can
request that it stop when the cell modem is disabled.

Make sure pppd is fully terminated by sending it SIGINT, SIGTERM, then
SIGKILL.

Add logic that prevents us from reacting if the the "cell_modem.enable"
setting is already "True".