Fix another potential crash, purge change detection code
Remove the change detection code and just hope that the pppd clean-up
code is sufficient enough to terminate an existing pppd daemon.  The
change detection code was trying to prevent launching multiple copies of
pppd if we'd already been set from an enabled value of False to True and
then received another message to set the enable value to True... which
seems like enough of a corner case to not be worth the complexity.

There was still a crash happening in async_output_cb, which could be
from uninitialized data, so initialize our stack buffer just to safe.