Refactor socket and loop to generic API
Creates two new API constructs in libpiksi:
 - Piksi Endpoint (replaces zsock_t usage)
 - Piksi Loop (replaces zloop_t usage)

The 'zmq' has been refactored out of the pubsub,
rx and tx contexts and whereas before a zloop
would come with the pubsub for free, there is now
an explicit 'attach' that must occur in order for
an rx context to have it's endpoint polled by the
loop.

Implements libuv usage in Piksi Loop API to
replace zloop with libuv's uv_loop_t, as well as
encapsulating the use of czmq completely in the
new Piksi Endpoint construct and associated APIs.

The sample_daemon package is refactored to use
the new APIs as an example of the sort of refactor
that will follow for the remaining packages.