Allow GGA upload to run continuously
Make sure resources get cleaned up properly.

Track consecutive read errors and cache the last valid GGA buffer so we
can send that when we occasionally can't open the file.

Since the file that caches the GGA information is occasioally not
available we only need to decide that we're in an error condition when
there's a number of consecutive errors.