Remove czmq style bind/connect indicator prefixes
In the czmq API, an '@' will bind and '>' will connect.
This has been replaced in code by explicitly creating
an endpoint of %type_name%_SERVER in the case of PUB/SUB
where globally in the Piksi all servers sockets are bound
in the router process. So this has simply been coded into
the router now, and all non-router PUB/SUB endpoint shall
use connect.