Latency fix: service all pending reads before exiting
The previous incarnation of pk_endpoint_receive had a loop that would
service all pending reads on a socket before exiting.  Bring back this
behavior to minimize the number of poll/select wake-ups that are
required to service packets.

This behavior is really important since the endpoint_router typically
operates on large numbers of small packets.  If we don't do this,
latencies start to grow into the seconds range as the number of packets
going through the system increases.