Improve error handling and start-up, trace asserts
+ On start-up when we're associating a server socket with the loop we
call pk_endpoint_loop_add twice, once to associate the endpoint with the
loop so we can wake up to accept new clients, and a second time for
*_SUB_SERVER and _*REP sockets so that we can read data from
clients.

Make sure we don't overwrite the accept handle with the wake-up-for-read
handle, this handle is only used when closing sockets, but could cause
us to disassociate the loop from the wrong handle.  It's unclear if we
need to keep the wake-up-for-read handle.

+ Add better error handling throughout and make the server
(endpoint_router) able to handle a restart by unlinking sockets prior to
creating them.

+ Attempt to make asserts more helpful by logging a backtrace of where
they fired.

+ Always shutdown and close a socket that's in an error state,
regardless of wether it's a client connection or not.