Add query_config util for reading config.ini
Use the minIni library to actually parse the config.ini instead of using
hacky regexes.

Both Dennis and I have seen device not boot in under certain
configurations.  The theory is that there's a certain arrangement of
config.ini data that causes the guards in the S85piksi_firmware script
to think that a port is enabled when it's not.

To remove this a possible problem, we add a `query_config` util which
uses the same minIni library to parse the config.ini and read the data
out.

This exetremely simple utility can we used in number of other place
later too.