Add test helpers, resolve duplicate message bug
The endpoint_router had a merge bug that was causing messages to be
duplicated during processing.  Created some tests to isolate where
the problem was.

The `endpoint_router` can be run locally (inside `make docker-shell` or
`make docker-exec`):

    LD_LIBRARY_PATH=/piksi_buildroot/buildroot/host_output/target/usr/lib \
      ./buildroot/host_output/target/usr/bin/endpoint_router --debug --name test --file \
      /piksi_buildroot/package/endpoint_router/endpoint_router/test/host_build.yml

Then, socat can be used to connect to the pub/sub sockets and manually
inspect the behavior.  To connect to a sub socket

    socat -u stdin UNIX-CONNECT:/tmp/sub,type=5

To see the data output from a pub socket:

    socat -u UNIX-CONNECT:/tmp/pub2,type=5 stdout

Similarly, the a pub/sub socket pair can be launched with the following
tester:

    LD_LIBRARY_PATH=/piksi_buildroot/buildroot/host_output/target/usr/lib \
      /piksi_buildroot/buildroot/host_output/build/libpiksi-0.1/test/pk_test_server \
      --pub ipc:///tmp/pub --sub ipc:///tmp/sub

The test server simply copies anything from the sub socket to the pub
socket, so you'd connect with socat as follows to see any data input
data from the pub socket arrive to the sub sockets:

    socat -u UNIX-CONNECT:/tmp/pub,type=5 stdout