Give resource_monitor it's own socket
Because resource_monitor doesn't current process any SBP messages it
doesn't need to be attached to the "internal" SBP pipe which sees
a lot of traffic.

Because there's a lot of traffic the resource_monitor gets behind on
receiving SBP data while it's inspecting the system and gets up
dropping data.