Use a pair of helper threads for CAN
Since CAN doesn't seem to work in O_NONBLOCK mode with libuv (we haven't
found the root cause of this yet).  Work around the issue by using a
pair of helper threads to read/write data from/to the SocketCAN handle.

The read thread does a "non-blocking" select to poll the descriptor to
see if it's ready, then sleeps for 1ms if it's not.  CAN data that's
read is written to the "write" end of the CAN read data pipe.

The write thread does a normal select on the read end of the write pipe
waiting for data from the main endpoint_adapter thread.  Data that
arrives is sent over CAN via a blocking write.