Non-forking endpoint adapter [ESD-103] (#1042)
## Design notes

* Make endpoint_adapter non-blocking:

To improve the efficiency of the endpoint adapter (and to clean up the process
table) we change to using an event loop from pk_loop_t in order to read and
service data in endpoint_adapter. This removes 3 processes per adapter in the
system, but leaves a "fork per client" for the TCP server. We can eventually
remove the fork from the TCP server as well.

Reading is interleaved by adding all serviced
sockets/endpoints/file-descriptors to an event loop.  Throughput favors
endpoints by using pk_endpoint_receive which services up to 32 reads from
endpoints (which is good, since endpoints typically have lots of small
messages).

* Use a pair of helper threads for CAN

Since CAN doesn't seem to work in O_NONBLOCK mode with libuv (we haven't found
the root cause of this yet).  Work around the issue by using a pair of helper
threads to read/write data from/to the SocketCAN handle.

The read thread does a "non-blocking" select to poll the descriptor to see if
it's ready, then sleeps for 1ms if it's not.  CAN data that's read is written
to the "write" end of the CAN read data pipe.

The write thread does a normal select on the read end of the write pipe waiting
for data from the main endpoint_adapter thread.  Data that arrives is sent over
CAN via a blocking write.

## Testing

Informal benchmarking using `watch -n1 'find
/var/log/metrics/sbp_fileio/external -type f -print -exec cat {} \;'` showed
maybe a slight slowdown with the non-blocking `endpoint_adapter` but the
different was only 2-5%.

The above write speed from sbp_fileio didn't seem to map to throughput from a
network perspective.  I was seeing about 200 kb/s of throughput, which visually
looks faster than our existing speed.