libpiksi, endpoint_adapter: inspect the last error [5/] [ESD-1038]
Add a means to inspect pk_endpoint_t and determine if the last error
that occured matches the given system error (this is done because libuv
records errors in a platform agnostic way, but we've tried to not leak
libuv things outside of libpiksi).

Use this new API to avoid a warning on EBADF, which happens when file
descriptors are closed, become invalid and are then kicked out of the
libuv event loop.