libpiksi, endpoint_adapter: sleep during send [6/] [ESD-1038]
These are tweaks to try to reduce dropped data caused by the router or
the adapters disconnecting a sockets (or dropping data) because it's
gotten behind on emptying it's queue.

Previously we'd disconnect or drop data in the case where a non-blocking
socket indicated that it would block (which means it's kernel buffer is
full).

Instead we now allow up to ~10ms of sleeping when sending data before
deciding to disconnect the socket or drop the data.