Interface tagging/filtering prototype
Allow data sent into the system to be tagged with bits of metadata and
use this metadata to filter out outgoing SBP messages according to
interface type.

This prototype introduces some overhead with send/recv in buffer copies,
a non-prototype implementation will probably introduce an API which
would force pk_endpoint_t users to allocate a buffer through a new
pk_endpoint_t call (which would allow pk_endpoint_t to reserve space
for the metadata).

We also match interfaces with strings on the incoming and outgoing
end of things.  This could be greatly improved by using a numeric ID
for the interface and checking against that rather than comparing
and copying strings for every SBP write message.