Close TCP listening socket in child processes (#1167)
After a call to fork() the newly spawned process receives a
copy of any open file descriptors, including listening sockets.
The child process doesn't require the listening socket, that is
exclusively for the use of the server process which just spawned
the child, so the child should close it immediately.

This change includes a slight rework of the forking functionality
in the general purpose io loop runner. The TCP server is the only
module which actually makes use of fork so it make sense to move
it out of the abstract function and put it in the place which
actually requires it. This way the server socket can be easily
close immediately after fork