Go back to sbp as a framer
The output filter for the SBP framer doesn't chop up outgoing messages
and assumes they are delivered as whole messages, which could cause
messages to drop of the first message matches an output filter.

Or, we could be writing too much data for the serial port to handle in
one go, which causes it to drop data.