Coldplug gets confused with newline characters
The coldplug script uses module aliases from sysfs to insert modules during
startup. It uses a series of shell commands piped in to each other with the
NULL separater options (-print0, -0, etc) presumably to cope with file names
containing spaces.

This causes problems when trying to parse a module alias which contains a newline
since it is using NULL as a separater characters. This fix changes the search
and probe command to a short loop which uses readline and newline characters