endpoint_adapter: don't drop data
We drop data if the remote client gets behind, it turns out that the
console can demonstrate how this can cause extremely bad behavior
in a client... when the console gets behind we'll start dropping
a few bytes from each send that arrives to endpoint_adapter (10-30
bytes).

Instead of doing this we'll just do our best in this situation and
report that the blocking is happening. In the case were we get really
behind (160k) the endpoint_router will eventually close the socket and
cause us to disconnect from the remote client.