scripts: build variant support infra [ESD-1085] [2/7]
Infrastructure for consuming the `build-variants.yaml` and
`external-artifacts.yaml` config files.

- `fetch-external-artifacts`: replaces the `fetch_firmware.sh` script
and uses the `external-artifacts.yaml` and `build-variant.yaml` file
to drive artifact fetching.

- `gen-docker-sync`: generates a docker-sync config based on the
DOCKER_HOST variant and the data in `build-variant.yaml`.

- `gen-requirements-yaml`: generates a `requiment.yaml` config for
consumption by HITL

- `gen-variant-configs`: a very simple script that takes in config
fragments listed in `build-variants.yaml` and dumps them into an
output file specified by the `config_output` config field.

- `pbr_script_lib.py`: library for code sharing between Python scripts

- `run-preflight-checks`: supports the `build-variants.yaml` concept of
"pre-flight" checks, which simply runs a set of scripts that verify the
build environment prior to launching a build

- `verify-aws-access`: makes sure AWS access is correctly configured
(this check is "smart", in that it won't actually check if the file
is already downloaded and the SHA256 matches the manifest).

- `verify-generated-configs`: makes sure that the configs in
`configs/*_defconfig` match their "generated" sources.

- `verify-piksi-ins-axx`: make sure that access to Piksi INS is
available if the variant is attempting to build INS support.

- `yaml2json`: support binaries for converting yaml to json script
Python does not have built-in support for YAML.