piksi-release variant support [ESD-1085] [6/7]
In order to make piksi_buildroot less dependent on piksi-releases in
order for a release to proceed, we introduce a `.piksi-releases.yaml`
config which allows the repository to specify a set of scripts that run
prior to the push of release commit that `piksi-releases` will create.
There is also support for a script which will list artifacts that should
be public (e.g. piksi_upgrade_tool).

In particular, the "post tag" script allows us to update the
`external-artifacts.yaml` file directly (with new version strings and
hashes) without the need for obfuscating mechanisms like m4 templates.