Add protocol blacklist to port mode selection
Certain CAN protocols such as J1939 are only applicable to the CAN ports
on the piksi. We should not be allowed to set the USB, UDP/TCP, etc.
ports into J1939 mode.

This commit accomplishes that by introducing protocol blacklists to the
ports daemon. Each port_config_t has a list of strings of protocol names
that are blacklisted on that port.

A list of all protocols is stored in a "global" array. Each port needs
to maintain their own list due to potential blacklisted protocols. Two
additional fields have been added to port_config_t to store the mapping
between these two lists.