Add protocol blacklist to port mode selection [ESD-1289] (#1274)
Certain CAN protocols such as J1939 are only applicable to the CAN ports
on the piksi. We should not be allowed to set the USB, UDP/TCP, etc.
ports into J1939 mode.

This commit accomplishes that by introducing protocol blacklists to the
ports daemon. Each port_config_t has a list of strings of protocol names
that are blacklisted on that port.

Previously, there was a one to one mapping between the list of protocols
we support and the mode enum values. Now that each port can have their
own blacklist, this mapping is neither valid nor consistent across
ports. We now maintain a mapping from mode to protocol index:
`port_config->mode_map[mode] = protocol_index`

Co-authored-by: Ben Altieri <benaltieri@gmail.com>